// src/app/interfaces/RTE_Sin_Comprobante.interface.ts
export interface sinComprobante {
    codigo_afiliado:               string;
    cuenta_afectada:               number;
    id_filial_ac:                  number;
    id_cliente:                    string;
    id_origen_fondos:              number;
    id_transaccion:                number;
    fecha_operacion:               string;
    monto_transaccion:             number;
    id_filial_realizo_transaccion: number;
    id_colaborador:                number;
    observaciones:                 string;
}